refactor(web): one card describes a session, wherever it is drawn
SessionFacts extracted verbatim from SessionRow's own body (title,
state, harness, model, task) so the collapsed rail's tooltip (next
commit) and the open row render the exact same card instead of two
implementations that drift -- the same argument rowMenu.ts makes
about verbs and task-reopen.ts makes about reopening.

VISUAL VERIFICATION PENDING (WSL port forwarding is down): that the
session list looks pixel-identical to before this extraction. tsc
and the full test suite are green, and the JSX moved is byte-for-byte
the same markup, just relocated into its own component.
