+++ packages/web/src/components/nav/SessionsAside.tsx
@@ -23,10 +23,10 @@ import {
 } from '@agentistics/tui/control/session-fleet'
 import { rowSelected } from '../../lib/fleetSelection'
 import { filterFleet, ignoredDimensions } from '../../lib/fleetFilter'
-import { HARNESS_COLORS, HARNESS_LABELS } from '../../lib/harness'
 import { NewSessionModal } from '../sessions/NewSessionModal'
 import { rowMenuEntries, type RowVerb } from '../../lib/rowMenu'
 import { SessionRowMenu } from '../sessions/SessionRowMenu'
+import { SessionFacts } from '../sessions/SessionFacts'
 import {
   MAX_PINNED, getPinnedIds, movePinnedSession, pinnedServerSnapshot, resolvePinnedRows,
   subscribePinnedSessions, togglePinnedSession,
@@ -124,18 +124,6 @@ const STATE_WASH: Record<string, string> = {
  * the moment somebody who pinned it wants it back. Where no conversation link can ever exist
  * (codex, kimi, gemini, agy — see `conversationBlind`) the row id is the only key there is.
  */
-/**
- * A model id, shortened for a narrow column.
- *
- * The provider prefix and the dated suffix are what a person already knows or does not care about
- * in a sidebar — `anthropic/claude-sonnet-4-5-20250929` becomes `claude-sonnet-4-5`. The full id is
- * on the row's `title` attribute, so nothing is lost.
- */
-function shortModel(model: string): string {
-  const bare = model.includes('/') ? model.slice(model.lastIndexOf('/') + 1) : model
-  return bare.replace(/-\d{8}$/, '')
-}
-
 function pinKeyOf(row: ControlSession): string {
   return row.conversationId ?? row.id
 }
@@ -731,45 +719,7 @@ function SessionRow({ session, selected, pinned, tap, onPin, onOpen, onMoveBy, v
           opacity: wants ? 1 : 0.55,
         }}
       />
-      <span style={{ minWidth: 0, flex: 1, display: 'flex', flexDirection: 'column', gap: 3 }}>
-        <span style={{
-          fontSize: 12.5, fontWeight: selected || wants ? 650 : 500,
-          overflow: 'hidden', textOverflow: 'ellipsis', whiteSpace: 'nowrap',
-        }}>
-          {session.title}
-        </span>
-        <span style={{
-          display: 'flex', alignItems: 'center', gap: 5, minWidth: 0,
-          fontSize: 10.5, color: wants ? 'var(--anthropic-orange)' : 'var(--text-tertiary)',
-        }}>
-          <span style={{ flexShrink: 0 }}>{session.stateLabel}</span>
-          <span style={{ opacity: 0.4, flexShrink: 0 }}>·</span>
-          <span style={{
-            color: (HARNESS_COLORS as Record<string, string>)[session.harness] ?? 'var(--text-tertiary)',
-            fontWeight: 650, flexShrink: 0,
-          }}>
-            {(HARNESS_LABELS as Record<string, string>)[session.harness] ?? session.harness}
-          </span>
-          {/* The model, when the row knows one. A row that does not is not "some default model" —
-              it is unknown, and inventing a name there is the confident-zero defect in words. */}
-          {session.model && (
-            <>
-              <span style={{ opacity: 0.4, flexShrink: 0 }}>·</span>
-              <span style={{ minWidth: 0, overflow: 'hidden', textOverflow: 'ellipsis', whiteSpace: 'nowrap' }}>
-                {shortModel(session.model)}
-              </span>
-            </>
-          )}
-          {session.task && (
-            <>
-              <span style={{ opacity: 0.4, flexShrink: 0 }}>·</span>
-              <span style={{ minWidth: 0, overflow: 'hidden', textOverflow: 'ellipsis', whiteSpace: 'nowrap' }}>
-                {session.task}
-              </span>
-            </>
-          )}
-        </span>
-      </span>
+      <SessionFacts session={session} selected={selected} />
       {/* The assistant, NAMED. It was a 5px dot, which carries the fact in colour alone — and a
           colour is not a name. The model sits with it on the meta line below. */}
       {/* The pin lives on the row rather than in a menu: it is a one-click decision about the row
